Extract helper function
